added mixpanel headers to /init api endpoint
