+++ .env.sample
@@ -13,3 +13,6 @@ REACT_APP_DEPLOYMENT_KEY=
 # Region for the host (i.e stage-us1, dev-us1, us1), defaults to us1(prod)
 REACT_APP_REGION=''
 
+# Webchat Version
+WEBCHAT_VERSION=''
+
